Add a blacklist with higher priority than whitelist
In cases where it is desireable to log objects with many keys, mostly
whitelisted but with a few exceptions nested deep in the object, the
whitelist can become large and repetitive. This is resolved by adding a
blacklist allowing wildcarding, allowing anonymization of single keys,
deeply nested.

For defensive purposes, the blacklist should be prioritized more highly
than the whitelist. The assumption here is that it is of great
importance to anonymize an item placed on the blacklist. This also
prevents the class of errors in which items with wildcards are added to
the whitelist after initial implementation, causing sensitive data to be
logged despite being present on the blacklist.

One side effect or gotcha is that keys that are a super string of blacklisted
items with wildcards will be redacted regardless of their presence on
the whitelist and not explicitely being thought of for blacklisting
(i.e. the motivation for adding `toAnonymize.*` is to recursively
redact values under the `toAnonymize` key rather than keys of the type
`toAnonymizeSuperString`).

https://github.com/uphold/anonymizer/issues/27